Move common to file
